+++ app/Models/OnlineTraining.php
@@ -4,6 +4,7 @@ namespace App\Models;
 
 use DB;
 use App\Models\Imageable;
+use Illuminate\Support\Str;
 use Illuminate\Database\Eloquent\Model;
 
 class OnlineTraining extends Model
@@ -79,7 +80,7 @@ class OnlineTraining extends Model
 
               // Image
               if(isset($value['base64Image'])) {
-                if($value['base64Image']) {
+                if($value['base64Image'] && !Str::contains($value['base64Image'], ['uploads'])) {
                   $image = Imageable::uploadImage($value['base64Image']);
                   $row->image()->delete();
                   $row->image()->create(['url' => $image]);
